+++ dags/modules/convert/pyarrow_to_trino_schema.py
@@ -1,4 +1,5 @@
 import logging
+import constants
 
 
 def pyarrow_to_trino_schema(schema):
